requirements.txt
click==8.0.0
colorama==0.4.6
greenlet==3.2.3
passlib==1.7.4
psycopg2==2.9.10
PyJWT==2.4.0
python-dotenv==1.1.0
pytest==7.4.3
pytest-cov==4.1.0
sentry-sdk==1.39.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
